+++ tardis/gui/__init__.py
@@ -1,2 +1,2 @@
 #Get the main-window into the namespace of the package
-from widgets import Tardis
+from interface import Tardis
