+++ tardis/gui/widgets.py
@@ -1,16 +1,4 @@
 import os
-from pkg_resources import parse_version
-import exceptions
-
-import numpy as np
-import matplotlib
-import matplotlib.pylab as plt
-import matplotlib.gridspec as gridspec
-from matplotlib import colors
-from matplotlib.patches import Circle
-from matplotlib.figure import *
-from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.backends.backend_qt4 import NavigationToolbar2QT as NavigationToolbar
 if os.environ.get('QT_API', None)=='pyqt':
     from PyQt4 import QtGui, QtCore
 elif os.environ.get('QT_API', None)=='pyside':
@@ -19,189 +7,173 @@ else:
     raise ImportError('QT_API was not set! Please exit the IPython console\n'
         ' and at the bash prompt use : \n\n export QT_API=pyside \n or\n'
         ' export QT_API=pyqt \n\n For more information refer to user guide.')
-import yaml
-from astropy import units as u
 
-from tardis import analysis, util
-from tardis import run_tardis
+import matplotlib
+from matplotlib.figure import *
+import matplotlib.gridspec as gridspec
+from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.backends.backend_qt4 import NavigationToolbar2QT as NavigationToolbar
+from matplotlib import colors
+from matplotlib.patches import Circle
+import matplotlib.pylab as plt
+
 import tardis
 
-if (parse_version(matplotlib.__version__) >= parse_version('1.4')):
-    matplotlib.style.use('fivethirtyeight')
-else:
-    print "Please upgrade matplotlib to a version >=1.4 for best results!"
-matplotlib.rcParams['font.family'] = 'serif'
-matplotlib.rcParams['font.size'] = 10.0
-matplotlib.rcParams['lines.linewidth'] = 1.0 
-matplotlib.rcParams['axes.formatter.use_mathtext'] = True
-matplotlib.rcParams['axes.edgecolor'] = matplotlib.rcParams['grid.color']
-matplotlib.rcParams['axes.linewidth'] = matplotlib.rcParams['grid.linewidth']
+class MatplotlibWidget(FigureCanvas):
+    """Canvas to draw graphs on."""
 
-class TemporarilyUnavaliable(Exception):
-    """Exception raised when creation of active mode of tardis is attempted."""
-    
-    def __init__(self, value):
-        self.value = value
-    
-    def __str__(self):
-        return repr(self.value)
+    def __init__(self, parent, fig=None):
+        """Create the canvas. Add toolbar depending on the parent."""
+        self.parent = parent
+        self.figure = Figure()#(frameon=False,facecolor=(1,1,1))
+        self.cid = {}
+        if fig != 'model':
+            self.ax = self.figure.add_subplot(111)
+        else:
+            self.gs = gridspec.GridSpec(2, 1, height_ratios=[1, 3])
+            self.ax1 = self.figure.add_subplot(self.gs[0])
+            self.ax2 = self.figure.add_subplot(self.gs[1])#, aspect='equal')
+        self.cb = None
+        self.span = None
 
-class Tardis(QtGui.QMainWindow):
-    """Create the top level window for the GUI and wait for call to 
-    display data.
+        super(MatplotlibWidget, self).__init__(self.figure)
+        super(MatplotlibWidget, self).setSizePolicy(QtGui.QSizePolicy.Expanding, 
+            QtGui.QSizePolicy.Expanding)
+        super(MatplotlibWidget, self).updateGeometry()
+        if fig != 'model':
+            self.toolbar = NavigationToolbar(self, parent)
+            self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
+            self.on_span_pick)
+        else:
+            self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
+                self.on_shell_pick)
+
+    def show_line_info(self):
+        """Show line info for span selected region."""
+        self.parent.line_info.append(LineInfo(self.parent, self.span.xy[0][0], 
+            self.span.xy[2][0]))
+
+    def show_span(self, garbage=0, left=5000, right=10000):
+        """Hide/Show/Change the buttons that show line info
+        in spectrum plot widget.
+
+        """
+        if self.parent.spectrum_span_button.text() == 'Show Wavelength Range':
+            if not self.span:
+                self.span = self.ax.axvspan(left, right, color='r', alpha=0.3, 
+                    picker=self.span_picker)
+            else:
+                self.span.set_visible(True)
+            self.parent.spectrum_line_info_button.show()
+            self.parent.spectrum_span_button.setText('Hide Wavelength Range')
+        else:
+            self.span.set_visible(False)
+            self.parent.spectrum_line_info_button.hide()
+            self.parent.spectrum_span_button.setText('Show Wavelength Range')
+        self.draw()
+
+    def on_span_pick(self, event):
+        """Callback to 'pick'(grab with mouse) the span selector tool."""
+        self.figure.canvas.mpl_disconnect(self.cid[0])
+        self.span.set_edgecolor('m')
+        self.span.set_linewidth(5)
+        self.draw()
+        if event.edge == 'left':
+            self.cid[1] = self.figure.canvas.mpl_connect('motion_notify_event', 
+                self.on_span_left_motion)
+        elif event.edge == 'right':
+            self.cid[1] = self.figure.canvas.mpl_connect('motion_notify_event', 
+                self.on_span_right_motion)
+        self.cid[2] = self.figure.canvas.mpl_connect('button_press_event', 
+            self.on_span_resized)
+
+    def on_span_left_motion(self, mouseevent):
+        """Update data of span selector tool on left movement of mouse and
+        redraw.
+
+        """
+        if mouseevent.xdata < self.span.xy[2][0]:
+            self.span.xy[0][0] = mouseevent.xdata
+            self.span.xy[1][0] = mouseevent.xdata
+            self.span.xy[4][0] = mouseevent.xdata
+            self.draw()
+
+    def on_span_right_motion(self, mouseevent):
+        """Update data of span selector tool on right movement of mouse and
+        redraw.
+
+        """
+        if mouseevent.xdata > self.span.xy[0][0]:
+            self.span.xy[2][0] = mouseevent.xdata
+            self.span.xy[3][0] = mouseevent.xdata
+            self.draw()
+
+    def on_span_resized(self, mouseevent):
+        """Redraw the red rectangle to currently selected span."""
+        self.figure.canvas.mpl_disconnect(self.cid[1])
+        self.figure.canvas.mpl_disconnect(self.cid[2])
+        self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
+            self.on_span_pick)
+        self.span.set_edgecolor('r')
+        self.span.set_linewidth(1)
+        self.draw()
+
+    def on_shell_pick(self, event):
+        """Highlight the shell that was picked."""
+        self.highlight_shell(event.artist.index)
+
+    def highlight_shell(self, index):
+        """Change edgecolor of highlighted shell."""
+        self.parent.tableview.selectRow(index)
+        for i in range(len(self.parent.shells)):
+            if i != index and i != index + 1:
+                self.parent.shells[i].set_edgecolor('k')
+            else:
+                self.parent.shells[i].set_edgecolor('w')
+        self.draw()
+
+    def shell_picker(self, shell, mouseevent):
+        """Enable picking shells in the shell plot."""
+        if mouseevent.xdata is None:
+            return False, dict()
+        mouse_r2 = mouseevent.xdata ** 2 + mouseevent.ydata ** 2
+        if shell.r_inner ** 2 < mouse_r2 < shell.r_outer ** 2:
+            return True, dict()
+        return False, dict()
+
+    def span_picker(self, span, mouseevent, tolerance=5):
+        """Detect mouseclicks inside tolerance region of the span selector
+        tool and pick it.
+
+        """
+        left = float(span.xy[0][0])
+        right = float(span.xy[2][0])
+        tolerance = span.axes.transData.inverted().transform((tolerance, 0)
+            )[0] - span.axes.transData.inverted().transform((0, 0))[0]
+        event_attributes = {'edge': None}
+        if mouseevent.xdata is None:
+            return False, event_attributes
+        if left - tolerance <= mouseevent.xdata <= left + tolerance:
+            event_attributes['edge'] = 'left'
+            return True, event_attributes
+        elif right - tolerance <= mouseevent.xdata <= right + tolerance:
+            event_attributes['edge'] = 'right'
+            return True, event_attributes
+        return False, event_attributes
+
+class Shell(matplotlib.patches.Wedge):
+    """A data holder to store measurements of shells that will be drawn in 
+    the plot.
 
     """
-
-    def __init__(self, parent=None, config=None, atom_data=None):
-        """Create the top level window and all widgets it contains.
-
-        When called with no arguments it initializes the GUI in passive
-        mode. When a yaml config file and atom data are provided the 
-        GUI starts in the active mode.
-
-        Parameters
-        ---------
-            parent: None
-                Set to None by default and shouldn't be changed unless 
-                you are developing something new.
-            config: string 
-                yaml file with configuration information for TARDIS.
-            atom_data: string
-                hdf file that has the atom data.
-
-        Raises
-        ------
-            TemporarilyUnavaliable
-                Raised when an attempt is made to start the active mode. 
-                This will be removed when active mode is developed.
-
-        """
-
-        #assumes that qt has already been initialized by starting IPython 
-        #with the flag "--pylab=qt"
-        app = QtCore.QCoreApplication.instance()
-        if app is None:
-            app = QtGui.QApplication([])
-        try:
-            from IPython.lib.guisupport import start_event_loop_qt4
-            start_event_loop_qt4(app)
-        except ImportError:
-            app.exec_()
-
-        super(Tardis, self).__init__(parent)
-
-        #path to icons folder
-        self.path = os.path.join(tardis.__path__[0],'gui','images')  
-
-        #Check if configuration file was provided
-        self.mode = 'passive'
-        if config is not None:
-            self.mode = 'active'
-
-        #Statusbar
-        statusbr = self.statusBar()
-        lblstr = '<font color="red"><b>Calculation did not converge</b></font>'
-        self.successLabel = QtGui.QLabel(lblstr)
-        self.successLabel.setFrameStyle(QtGui.QFrame.StyledPanel |
-            QtGui.QFrame.Sunken)
-        statusbr.addPermanentWidget(self.successLabel)
-        self.modeLabel = QtGui.QLabel('Passive mode')
-        statusbr.addPermanentWidget(self.modeLabel)
-        statusbr.showMessage(self.mode, 5000)
-        statusbr.showMessage("Ready", 5000) 
-
-        #Actions
-        quitAction = QtGui.QAction("&Quit", self)
-        quitAction.setIcon(QtGui.QIcon(os.path.join(self.path, 
-            'closeicon.png')))
-        quitAction.triggered.connect(self.close)
-        
-        self.viewMdv = QtGui.QAction("View &Model", self)
-        self.viewMdv.setIcon(QtGui.QIcon(os.path.join(self.path,
-            'mdvswitch.png')))
-        self.viewMdv.setCheckable(True)
-        self.viewMdv.setChecked(True)
-        self.viewMdv.setEnabled(False)
-        self.viewMdv.triggered.connect(self.switchToMdv)
-        
-        self.viewForm = QtGui.QAction("&Edit Model", self)
-        self.viewForm.setIcon(QtGui.QIcon(os.path.join(self.path,
-            'formswitch.png')))
-        self.viewForm.setCheckable(True)
-        self.viewForm.setEnabled(False)
-        self.viewForm.triggered.connect(self.switchToForm)
-
-        #Menubar
-        self.fileMenu = self.menuBar().addMenu("&File")
-        self.fileMenu.addAction(quitAction)
-        self.viewMenu = self.menuBar().addMenu("&View")
-        self.viewMenu.addAction(self.viewMdv)
-        self.viewMenu.addAction(self.viewForm)
-        self.helpMenu = self.menuBar().addMenu("&Help")
-
-        #Toolbar
-        fileToolbar = self.addToolBar("File")
-        fileToolbar.setObjectName("FileToolBar")  
-        fileToolbar.addAction(quitAction)
-
-        viewToolbar = self.addToolBar("View")
-        viewToolbar.setObjectName("ViewToolBar")
-        viewToolbar.addAction(self.viewMdv)
-        viewToolbar.addAction(self.viewForm)
-
-        #Central Widget
-        self.stackedWidget = QtGui.QStackedWidget()
-        self.mdv = ModelViewer() 
-        self.stackedWidget.addWidget(self.mdv)
-        
-        #In case of active mode
-        if self.mode == 'active':
-            #Disabled currently
-            # self.formWidget = ConfigEditor(config)
-            # #scrollarea
-            # scrollarea = QtGui.QScrollArea()
-            # scrollarea.setWidget(self.formWidget)
-            # self.stackedWidget.addWidget(scrollarea)
-            # self.viewForm.setEnabled(True)
-            # self.viewMdv.setEnabled(True)
-            # model = run_tardis(config, atom_data)
-            # self.show_model(model)
-            raise TemporarilyUnavaliable("The active mode is under" 
-                "development. Please use the passive mode for now.")
-
-        self.setCentralWidget(self.stackedWidget)
-
-    def show_model(self, model=None):
-        """Set the provided model into the GUI and show the main window.
-
-        Parameters
-        ----------
-        model: TARDIS model object 
-            A keyword argument that takes the tardis model object.
-
-        """
-        if model:
-            self.mdv.change_model(model)
-        if model.converged:
-            self.successLabel.setText('<font color="green">converged</font>')
-        if self.mode == 'active':
-            self.modeLabel.setText('Active Mode')
-
-        self.mdv.fillOutputLabel()
-        self.mdv.tableview.setModel(self.mdv.tablemodel)
-        self.mdv.plot_model()
-        self.mdv.plot_spectrum()
-        self.showMaximized()
-
-    def switchToMdv(self):
-        """Switch the cental stacked widget to show the modelviewer."""
-        self.stackedWidget.setCurrentIndex(0)
-        self.viewForm.setChecked(False)
-
-    def switchToForm(self):
-        """Switch the cental stacked widget to show the ConfigEditor."""
-        self.stackedWidget.setCurrentIndex(1)
-        self.viewMdv.setChecked(False)
+    def __init__(self, index, center, r_inner, r_outer, **kwargs):
+        super(Shell, self).__init__(center, r_outer, 0, 90, 
+            width=r_outer - r_inner, **kwargs)
+        self.index = index
+        self.center = center
+        self.r_outer = r_outer
+        self.r_inner = r_inner
+        self.width = r_outer - r_inner
 
 class ConfigEditor(QtGui.QWidget):
     """The configuration editor widget. 
@@ -412,459 +384,6 @@ class ConfigEditor(QtGui.QWidget):
         """
         pass
 
-class Node(object):
-    """Object that serves as the nodes in the TreeModel.
-
-    Attributes
-    ----------
-        parent: None/Node 
-            The parent of the node.
-        children: list of Node
-            The children of the node.
-        data: list of string 
-            The data stored on the node. Can be a key or a value.
-        siblings: dictionary 
-            A dictionary of nodes that are siblings of this node. The 
-            keys are the values of the nodes themselves. This is 
-            used to keep track of which value the user has selected
-            if the parent of this node happens to be a key that can
-            take values from a list.
-
-    """
-
-    def __init__(self, data, parent=None):
-        """Create one node with the data and parent provided.
-
-        Parameters
-        ----------
-            data: list of string
-                The data that is intended to be stored on the node.
-            parent: Node 
-                Another node which is the parent of this node. The 
-                root node has parent set to None.
-
-        Note
-        ----
-            A leaf node is a node that is the only child of its parent. 
-            For this tree this will always be the case. This is because
-            the tree stores the key at every node except the leaf node
-            where it stores the value for the key. So if in the dictionary
-            the value of a key is another dictionary, then it will 
-            be a node with no leafs. If the key has a value that is a 
-            value or a list then it will have one child that is a leaf.
-            The leaf can have no children. For example-
-
-            'a':val1
-            'b':{'c':val2
-                 'd':val3
-                 'e':{'f':val4
-                      'g':val5
-                         :val6
-                         :val7}} *In this case the key g can take values
-                                  val7, val5 and val6 and is currently
-                                  set to val5.
-
-            In the tree shown above all quoted values are keys in the 
-            dictionary and are non-leaf nodes in the tree. All values 
-            of the form valx are leaf nodes and are not dictionaries 
-            themselves. If the keys have non-dictionary values then they
-            have a leaf attached. And no leaf can have a child.
-        
-        """
-        self.parent = parent
-        self.children = []
-        self.data = data
-        self.siblings = {}  #For 'type' fields. Will store the nodes to 
-                            #enable disable on selection
-
-    def appendChild(self, child):
-        """Add a child to this node."""
-        self.children.append(child)
-        child.parent = self
-
-    def getChild(self, i):
-        """Get the ith child of this node. 
-
-        No error is raised if the cild requested doesn't exist. A 
-        None is returned in such cases.
-
-        """
-        if i < self.numChildren():
-            return self.children[i]
-        else:
-            return None
-
-    def numChildren(self):
-        """Number of children this node has."""
-        return len(self.children)
-
-    def numColumns(self):
-        """Returns the number of strings stored in the data attribute."""
-        return len(self.data)
-
-    def getData(self, i):
-        """Returns the ith string from the data list.
-
-        No error is raised if the data list index is exceeded. A None is
-        returned in such cases.
-
-        """
-        try:
-            return self.data[i]
-        except IndexError:
-            return None
-
-    def getParent(self):
-        """Return the parent of this node."""
-        return self.parent
-
-    def getIndexOfSelf(self):
-        """Returns the number at which it comes in the list of its 
-        parent's children. For root the index 0 is returned.
-
-        """
-        if self.parent:
-            return self.parent.children.index(self)
-        else:
-            return 0
-
-    def setData(self, column, value):
-        """Set the data for the ith index to the provided value. Returns
-        true if the data was set successfully.
-
-        """
-        if column < 0 or column >= self.numColumns():
-            return False
-
-        self.data[column] = value
-
-        return True
-
-class TreeModel(QtCore.QAbstractItemModel):
-    """The class that defines the tree for ConfigEditor.
-
-    Parameters
-    ----------
-        root: Node
-            Root node of the tree.
-        disabledNodes: list of Node 
-            List of leaf nodes that are not editable currently.
-        typenodes: list of Node 
-            List of nodes that correspond to keys that set container 
-            types. Look at tardis configuration template. These are the
-            nodes that have values that can be set from a list.
-
-    """
-    def __init__(self, dictionary, parent=None):
-        """Create a tree of tardis configuration dictionary.
-
-        Parameters
-        ----------
-            dictionary: dictionary            
-                The dictionary that needs to be converted to the tree.
-            parent: None 
-                Used to instantiate the QAbstractItemModel
-
-        """
-        QtCore.QAbstractItemModel.__init__(self, parent)
-
-        self.root = Node(["column A"])
-        self.disabledNodes = []
-        self.typenodes = []
-        self.dictToTree(dictionary, self.root)
-
-    def dictToTree(self, dictionary, root):
-        """Create the tree and append siblings to nodes that need them.
-
-        Parameters
-        ----------
-            dictionary: dictionary
-                The dictionary that is to be converted to the tree.
-            root: Node 
-                The root node of the tree.
-
-        """
-        #Construct tree with all nodes
-        self.treeFromNode(dictionary, root)
-
-        #Append siblings to type nodes
-        for node in self.typenodes: #For every type node
-            parent = node.getParent()
-            sibsdict = {}
-            for i in range(parent.numChildren()):
-                sibsdict[parent.getChild(i).getData(0)] = parent.getChild(i)
-
-            typesleaf = node.getChild(0)
-            for i in range(typesleaf.numColumns()):
-                sibstrings = typesleaf.getData(i).split('|_:_|')
-            
-                typesleaf.setData(i, sibstrings[0])
-                sibslist = []
-                for j in range(1, len(sibstrings)):
-                    if sibstrings[j] in sibsdict:
-                        sibslist.append(sibsdict[sibstrings[j]])
-
-                typesleaf.siblings[sibstrings[0]] = sibslist
-            
-            #Then append siblings of current selection for all type nodes to
-            #disabled nodes
-            for i in range(1,typesleaf.numColumns()):
-                key = typesleaf.getData(i)
-                for nd in typesleaf.siblings[key]:
-                    self.disabledNodes.append(nd)
-
-
-    def treeFromNode(self, dictionary, root):
-        """Convert dictionary to tree. Called by dictToTree."""
-        for key in dictionary:
-            child = Node([key])
-            root.appendChild(child)
-            if isinstance(dictionary[key], dict):
-                self.treeFromNode(dictionary[key], child)
-            elif isinstance(dictionary[key], list):
-                if isinstance(dictionary[key][1], list):
-                    leaf = Node(dictionary[key][1])    
-                else:
-                    leaf = Node([dictionary[key][1]])
-
-                child.appendChild(leaf)
-                if key == 'type':
-                    self.typenodes.append(child)
-
-    def dictFromNode(self, node): 
-        """Take a node and convert the whole subtree rooted at it into a 
-        dictionary.
-
-        """
-        children = [node.getChild(i) for i in range(node.numChildren())]
-        if len(children) > 1:
-            dictionary = {}
-            for nd in children:
-                if nd in self.disabledNodes:
-                    pass
-                else:
-                    dictionary[nd.getData(0)] = self.dictFromNode(nd)
-            return dictionary
-        elif len(children)==1:
-            return children[0].getData(0)
-
-    def columnCount(self, index):
-        """Return the number of columns in the node pointed to by
-        the given model index.
-
-        """
-        if index.isValid():
-            return index.internalPointer().numColumns()
-        else:
-            return self.root.numColumns()
-
-    def data(self, index, role):
-        """Returns the asked data for the node specified by the modeLabel
-        index."""
-        if not index.isValid():
-            return None
-
-        if role != QtCore.Qt.DisplayRole:
-            return None
-
-        item = index.internalPointer()
-
-        return item.getData(index.column())
-
-    def flags(self, index):
-        """Return flags for the items whose model index is provided."""
-        if not index.isValid():
-            return QtCore.Qt.NoItemFlags
-
-        node = index.internalPointer()
-        if ((node.getParent() in self.disabledNodes) or 
-            (node in self.disabledNodes)):
-            return QtCore.Qt.NoItemFlags
-
-        if node.numChildren()==0:
-            return (QtCore.Qt.ItemIsEditable | QtCore.Qt.ItemIsEnabled | 
-                QtCore.Qt.ItemIsSelectable)
-
-        return QtCore.Qt.ItemIsEnabled | QtCore.Qt.ItemIsSelectable        
-
-    def getItem(self, index):
-        """Returns the node to which the model index is pointing. If the
-        model index is invalid then the root node is returned.
-
-        """
-        if index.isValid():
-            item = index.internalPointer()
-            if item:
-                return item
-
-        return self.root
-
-    def headerData(self, section, orientation, role):
-        """Returns header data. This is not used in QColumnView. But will
-        be needed for QTreeView.
-
-        """
-        if (orientation == QtCore.Qt.Horizontal and 
-          role == QtCore.Qt.DisplayRole):
-            return self.root.getData(section)
-
-        return None
-
-    def index(self, row, column, parent=QtCore.QModelIndex()):
-        """Create a model index for the given row and column. For a 
-        tree model, the row is the set of nodes with the same parents and
-        the column indexes the data in the node.
-
-        """
-        if parent.isValid() and parent.column() != 0:
-            return QtCore.QModelIndex()
-
-        parentItem = self.getItem(parent)
-        childItem = parentItem.getChild(row)
-        if childItem:
-            return self.createIndex(row, column, childItem)
-        else:
-            return QtCore.QModelIndex()
-
-    def insertColumns(self, position, columns, parent=QtCore.QModelIndex()):
-        """Insert columns in the tree model."""
-        self.beginInsertColumns(parent, position, position + columns - 1)
-        success = self.root.insertColumns(position, columns)
-        self.endInsertColumns()
-
-        return success
-
-    def insertRows(self, position, rows, parent=QtCore.QModelIndex()):
-        """Insert rows in the tree model."""
-        parentItem = self.getItem(parent)
-        self.beginInsertRows(parent, position, position + rows - 1)
-        success = parentItem.insertChildren(position, rows,
-                    self.rootItem.columnCount())
-        self.endInsertRows()
-
-        return success
-
-    def parent(self, index):
-        """Return the parent of the node to which the index points."""
-        if not index.isValid():
-            return QtCore.QModelIndex()
-
-        childItem = index.internalPointer()
-        parentItem = childItem.getParent()
-
-        if parentItem == self.root:
-            return QtCore.QModelIndex()
-
-        return self.createIndex(parentItem.getIndexOfSelf(), 0, parentItem)
-
-    def rowCount(self, parent=QtCore.QModelIndex()):
-        """The number of rows for a given node. 
-        
-        (The number of rows is just the number of children for a node.)
-
-        """
-        parentItem = self.getItem(parent)
-
-        return parentItem.numChildren()
-
-    def setData(self, index, value, role=QtCore.Qt.EditRole):
-        """Set the value as the data at the location pointed by the 
-        index.
-
-        """
-        if role != QtCore.Qt.EditRole:
-            return False
-
-        item = self.getItem(index)
-        result = item.setData(index.column(), value)
-
-        if result:
-            self.dataChanged.emit(index, index)
-            #print self.dictFromNode(self.root)
-        return result
-
-    def setHeaderData(self, section, orientation, value, 
-        role=QtCore.Qt.EditRole):
-        """Change header data. Unused in columnview."""
-        if role != QtCore.Qt.EditRole or orientation != QtCore.Qt.Horizontal:
-            return False
-
-        result = self.root.setData(section, value)
-        if result:
-            self.headerDataChanged.emit(orientation, section, section)
-
-        return result
-
-class TreeDelegate(QtGui.QStyledItemDelegate):
-    """Create a custom delegate to modify the columnview that displays the 
-    TreeModel.
-
-    """
-    def __init__(self, parent=None):
-        """Call the constructor of the superclass."""
-        QtGui.QStyledItemDelegate.__init__(self, parent)
-
-    def createEditor(self, parent, option, index):
-        """Create a lineEdit or combobox depending on the type of node."""
-        node = index.internalPointer()
-        if node.numColumns()>1:
-            combobox = QtGui.QComboBox(parent)
-            combobox.addItems([node.getData(i) for i in range(node.numColumns())])
-            combobox.setEditable(False)
-            return combobox
-        else:
-            editor =  QtGui.QLineEdit(parent)
-            editor.setText(str(node.getData(0)))
-            editor.returnPressed.connect(self.closeAndCommit)
-            return editor
-
-    def closeAndCommit(self):
-        """Saver for the line edits."""
-        editor = self.sender()
-        if isinstance(editor, QtGui.QLineEdit):
-            self.commitData.emit(editor)
-            self.closeEditor.emit(editor, QtGui.QAbstractItemDelegate.NoHint)
-
-    def setModelData(self, editor, model, index):
-        """Called when new data id set in the model. This is where the
-        siblings of type nodes are enabled or disabled according to the 
-        new choice made.
-
-        """
-        node = index.internalPointer()
-
-        if node.numColumns() > 1 and node.getParent().getData(0) != 'type':
-            selectedIndex = editor.currentIndex()
-            firstItem = node.getData(0)
-            node.setData(0, str(editor.currentText()))
-            node.setData(selectedIndex, str(firstItem))
-
-        elif node.numColumns() > 1 and node.getParent().getData(0) == 'type':
-            selectedIndex = editor.currentIndex()
-            firstItem = node.getData(0)
-            node.setData(0, str(editor.currentText()))
-            node.setData(selectedIndex, str(firstItem))
-
-            itemsToDisable = node.siblings[firstItem]
-            itemsToEnable = node.siblings[str(editor.currentText())]
-
-            for nd in itemsToDisable:
-                model.disabledNodes.append(nd)
-
-            for nd in itemsToEnable:
-                if nd in model.disabledNodes:
-                    model.disabledNodes.remove(nd) 
-
-        elif isinstance(editor, QtGui.QLineEdit): 
-            node.setData(0, str(editor.text()))
-        else:
-            QtGui.QStyledItemDelegate.setModelData(self, editor, model, index)
-            
-        #print model.dictFromNode(model.root) 
-        #f = open('dictester.dat','w')
-        #f.write(yaml.dump(model.dictFromNode(model.root)))
-        #f.close()
-
 class ModelViewer(QtGui.QWidget):
     """The widget that holds all the plots and tables that visualize 
     the data in the tardis model. This is also appended to the stacked 
@@ -872,14 +391,17 @@ class ModelViewer(QtGui.QWidget):
 
     """
     
-    def __init__(self, parent=None):
+    def __init__(self, tablecreator, parent=None):
         """Create all widgets that are children of ModelViewer."""
-        super(ModelViewer, self).__init__(parent)
+        QtGui.QWidget.__init__(self, parent)
         
         #Data structures
         self.model = None
         self.shell_info = {}
         self.line_info = []
+
+        #functions
+        self.createTable = tablecreator
 
         #Shells widget
         self.shellWidget = self.makeShellWidget()
@@ -893,8 +415,8 @@ class ModelViewer(QtGui.QWidget):
         self.plotTabWidget.addTab(self.spectrumWidget, "S&pectrum")
 
         #Table widget
-        self.tablemodel = SimpleTableModel([['Shell: '], ["Rad. temp", "Ws"]],
-         (1, 0))
+        self.tablemodel = self.createTable([['Shell: '], ["Rad. temp", "Ws"]],
+                            (1, 0))
         self.tableview = QtGui.QTableView()
         self.tableview.setMinimumWidth(200)
         self.tableview.connect(self.tableview.verticalHeader(), 
@@ -1021,7 +543,6 @@ class ModelViewer(QtGui.QWidget):
         containerWidget = QtGui.QWidget()
         containerWidget.setLayout(self.spectrum_sublayout)
         return containerWidget
-
 
     def update_data(self, model=None):
         """Associate the given model with the GUI and display results."""
@@ -1181,14 +702,16 @@ class ModelViewer(QtGui.QWidget):
 
     def on_header_double_clicked(self, index):
         """Callback to get counts for different Z from table."""
-        self.shell_info[index] = ShellInfo(index, self)
+        self.shell_info[index] = ShellInfo(index, self.createTable, self)
 
 class ShellInfo(QtGui.QDialog):
     """Dialog to display Shell abundances."""
 
-    def __init__(self, index, parent=None):
+    def __init__(self, index, tablecreator, parent=None):
         """Create the widget to display shell info and set data."""
         super(ShellInfo, self).__init__(parent)
+
+        self.createTable = tablecreator
         self.parent = parent
         self.shell_index = index
         self.setGeometry(400, 150, 200, 400)
@@ -1203,7 +726,7 @@ class ShellInfo(QtGui.QDialog):
 
         self.table1_data = self.parent.model.tardis_config.abundances[
             self.shell_index]
-        self.atomsdata = SimpleTableModel([['Z = '], ['Count (Shell %d)' % (
+        self.atomsdata = self.createTable([['Z = '], ['Count (Shell %d)' % (
             self.shell_index + 1)]], iterate_header=(2, 0), 
             index_info=self.table1_data.index.values.tolist())
         self.ionsdata = None
@@ -1226,7 +749,7 @@ class ShellInfo(QtGui.QDialog):
         self.current_atom_index = self.table1_data.index.values.tolist()[index]
         self.table2_data = self.parent.model.plasma_array.ion_populations[
             self.shell_index].ix[self.current_atom_index]
-        self.ionsdata = SimpleTableModel([['Ion: '], 
+        self.ionsdata = self.createTable([['Ion: '], 
             ['Count (Z = %d)' % self.current_atom_index]], 
             iterate_header=(2, 0), 
             index_info=self.table2_data.index.values.tolist())
@@ -1252,7 +775,7 @@ class ShellInfo(QtGui.QDialog):
         self.current_ion_index = self.table2_data.index.values.tolist()[index]
         self.table3_data = self.parent.model.plasma_array.level_populations[
             self.shell_index].ix[self.current_atom_index, self.current_ion_index]
-        self.levelsdata = SimpleTableModel([['Level: '], 
+        self.levelsdata = self.createTable([['Level: '], 
             ['Count (Ion %d)' % self.current_ion_index]], 
             iterate_header=(2, 0), 
             index_info=self.table3_data.index.values.tolist())
@@ -1279,85 +802,6 @@ class ShellInfo(QtGui.QDialog):
         self.levelstable.hide()
         self.setGeometry(400, 150, 200, 400)
         self.show()
-
-class LineInteractionTables(QtGui.QWidget):
-    """Widget to hold the line interaction tables used by 
-    LineInfo which in turn is used by spectrum widget.
-
-    """
-
-    def __init__(self, line_interaction_analysis, atom_data, description):
-        """Create the widget and set data."""
-        super(LineInteractionTables, self).__init__()
-        self.text_description = QtGui.QLabel(str(description))
-        self.species_table = QtGui.QTableView()
-        self.transitions_table = QtGui.QTableView()
-        self.layout = QtGui.QHBoxLayout()
-        self.line_interaction_analysis = line_interaction_analysis
-        self.atom_data = atom_data
-        line_interaction_species_group = \
-        line_interaction_analysis.last_line_in.groupby(['atomic_number', 
-            'ion_number'])
-        self.species_selected = sorted(
-            line_interaction_species_group.groups.keys())
-        species_symbols = [util.species_tuple_to_string(item, 
-            atom_data) for item in self.species_selected]
-        species_table_model = SimpleTableModel([species_symbols, ['Species']])
-        species_abundances = (
-            line_interaction_species_group.wavelength.count().astype(float) /
-            line_interaction_analysis.last_line_in.wavelength.count()).astype(float).tolist()
-        species_abundances = map(float, species_abundances)
-        species_table_model.addData(species_abundances)
-        self.species_table.setModel(species_table_model)
-
-        line_interaction_species_group.wavelength.count()
-        self.layout.addWidget(self.text_description)
-        self.layout.addWidget(self.species_table)
-        self.species_table.connect(self.species_table.verticalHeader(), 
-            QtCore.SIGNAL('sectionClicked(int)'), self.on_species_clicked)
-        self.layout.addWidget(self.transitions_table)
-
-        self.setLayout(self.layout)
-        self.show()
-
-    def on_species_clicked(self, index):
-        """"""
-        current_species = self.species_selected[index]
-        last_line_in = self.line_interaction_analysis.last_line_in
-        last_line_out = self.line_interaction_analysis.last_line_out
-
-        last_line_in_filter = (last_line_in.atomic_number == current_species[0]).values & \
-                              (last_line_in.ion_number == current_species[1]).values
-
-        current_last_line_in = last_line_in[last_line_in_filter].reset_index()
-        current_last_line_out = last_line_out[last_line_in_filter].reset_index()
-
-        current_last_line_in['line_id_out'] = current_last_line_out['line_id']
-
-
-        last_line_in_string = []
-        last_line_count = []
-        grouped_line_interactions = current_last_line_in.groupby(['line_id', 
-            'line_id_out'])
-        exc_deexc_string = 'exc. %d-%d (%.2f A) de-exc. %d-%d (%.2f A)'
-
-        for line_id, row in grouped_line_interactions.wavelength.count().iteritems():
-            current_line_in = self.atom_data.lines.ix[line_id[0]]
-            current_line_out = self.atom_data.lines.ix[line_id[1]]
-            last_line_in_string.append(exc_deexc_string % (
-                current_line_in['level_number_lower'],
-               current_line_in['level_number_upper'],
-               current_line_in['wavelength'],
-               current_line_out['level_number_upper'],
-               current_line_out['level_number_lower'],
-               current_line_out['wavelength']))
-            last_line_count.append(int(row))
-
-
-        last_line_in_model = SimpleTableModel([last_line_in_string, [
-            'Num. pkts %d' % current_last_line_in.wavelength.count()]])
-        last_line_in_model.addData(last_line_count)
-        self.transitions_table.setModel(last_line_in_model)
 
 class LineInfo(QtGui.QDialog):
     """Dialog to show the line info used by spectrum widget."""
@@ -1493,234 +937,248 @@ class LineInfo(QtGui.QDialog):
         self.setGeometry(180 + len(self.parent.line_info) * 20, 150, 750, 400)
         self.show()
 
-class SimpleTableModel(QtCore.QAbstractTableModel):
-    """Create a table data structure for the table widgets."""
-    
-    def __init__(self, headerdata=None, iterate_header=(0, 0), 
-            index_info=None, parent=None, *args):
-        """Call constructor of the QAbstractTableModel and set parameters
-        given by user.
-        """
-        super(SimpleTableModel, self).__init__(parent, *args)
-        self.headerdata = headerdata
-        self.arraydata = []
-        self.iterate_header = iterate_header
-        self.index_info = index_info
-
-    def addData(self, datain):
-        """Add data to the model."""
-        self.arraydata.append(datain)
-
-    def rowCount(self, parent=QtCore.QModelIndex()):
-        """Return number of rows."""
-        return len(self.arraydata[0])
-
-    def columnCount(self, parent=QtCore.QModelIndex()):
-        """Return number of columns."""
-        return len(self.arraydata)
-
-    def headerData(self, section, orientation, role=QtCore.Qt.DisplayRole):
-        """Set the header data."""
-        if orientation == QtCore.Qt.Vertical and role == QtCore.Qt.DisplayRole:
-            if self.iterate_header[0] == 1:
-                return self.headerdata[0][0] + str(section + 1)
-            elif self.iterate_header[0] == 2:
-                if self.index_info:
-                    return self.headerdata[0][0] + str(self.index_info[section])
-                else:
-                    return self.headerdata[0][0] + str(section + 1)
-            else:
-                return self.headerdata[0][section]
-        elif orientation == QtCore.Qt.Horizontal and role == QtCore.Qt.DisplayRole:
-            if self.iterate_header[1] == 1:
-                return self.headerdata[1][0] + str(section + 1)
-            elif self.iterate_header[1] == 2:
-                if self.index_info:
-                    return self.headerdata[1][0] + str(self.index_info[section])
-            else:
-                return self.headerdata[1][section]
-        return None
-
-    def data(self, index, role=QtCore.Qt.DisplayRole):
-        """Return data of specified index and role."""
-        if not index.isValid():
-            return None
-        elif role != QtCore.Qt.DisplayRole:
-            return None
-        return (self.arraydata[index.column()][index.row()])
-
-    def setData(self, index, value, role=QtCore.Qt.EditRole):
-        """Change the data in the model for specified index and role
-        to specified value."""
-        if not index.isValid():
-            return False
-        elif role != QtCore.Qt.EditRole:
-            return False
-        self.arraydata[index.column()][index.row()] = value
-        self.emit(QtCore.SIGNAL(
-            'dataChanged(const QModelIndex &, const QModelIndex &)'), 
-            index, index)
-        return True
-
-    def updateTable(self):
-        """Update table to set all the new data."""
-        for r in range(self.rowCount()):
-            for c in range(self.columnCount()):
-                index = self.createIndex(r, c)
-                self.setData(index, self.arraydata[c][r])
-
-class MatplotlibWidget(FigureCanvas):
-    """Canvas to draw graphs on."""
-
-    def __init__(self, parent, fig=None):
-        """Create the canvas. Add toolbar depending on the parent."""
-        self.parent = parent
-        self.figure = Figure()#(frameon=False,facecolor=(1,1,1))
-        self.cid = {}
-        if fig != 'model':
-            self.ax = self.figure.add_subplot(111)
-        else:
-            self.gs = gridspec.GridSpec(2, 1, height_ratios=[1, 3])
-            self.ax1 = self.figure.add_subplot(self.gs[0])
-            self.ax2 = self.figure.add_subplot(self.gs[1])#, aspect='equal')
-        self.cb = None
-        self.span = None
-
-        super(MatplotlibWidget, self).__init__(self.figure)
-        super(MatplotlibWidget, self).setSizePolicy(QtGui.QSizePolicy.Expanding, 
-            QtGui.QSizePolicy.Expanding)
-        super(MatplotlibWidget, self).updateGeometry()
-        if fig != 'model':
-            self.toolbar = NavigationToolbar(self, parent)
-            self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
-            self.on_span_pick)
-        else:
-            self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
-                self.on_shell_pick)
-
-    def show_line_info(self):
-        """Show line info for span selected region."""
-        self.parent.line_info.append(LineInfo(self.parent, self.span.xy[0][0], 
-            self.span.xy[2][0]))
-
-    def show_span(self, garbage=0, left=5000, right=10000):
-        """Hide/Show/Change the buttons that show line info
-        in spectrum plot widget.
-
-        """
-        if self.parent.spectrum_span_button.text() == 'Show Wavelength Range':
-            if not self.span:
-                self.span = self.ax.axvspan(left, right, color='r', alpha=0.3, 
-                    picker=self.span_picker)
-            else:
-                self.span.set_visible(True)
-            self.parent.spectrum_line_info_button.show()
-            self.parent.spectrum_span_button.setText('Hide Wavelength Range')
-        else:
-            self.span.set_visible(False)
-            self.parent.spectrum_line_info_button.hide()
-            self.parent.spectrum_span_button.setText('Show Wavelength Range')
-        self.draw()
-
-    def on_span_pick(self, event):
-        """Callback to 'pick'(grab with mouse) the span selector tool."""
-        self.figure.canvas.mpl_disconnect(self.cid[0])
-        self.span.set_edgecolor('m')
-        self.span.set_linewidth(5)
-        self.draw()
-        if event.edge == 'left':
-            self.cid[1] = self.figure.canvas.mpl_connect('motion_notify_event', 
-                self.on_span_left_motion)
-        elif event.edge == 'right':
-            self.cid[1] = self.figure.canvas.mpl_connect('motion_notify_event', 
-                self.on_span_right_motion)
-        self.cid[2] = self.figure.canvas.mpl_connect('button_press_event', 
-            self.on_span_resized)
-
-    def on_span_left_motion(self, mouseevent):
-        """Update data of span selector tool on left movement of mouse and
-        redraw.
-
-        """
-        if mouseevent.xdata < self.span.xy[2][0]:
-            self.span.xy[0][0] = mouseevent.xdata
-            self.span.xy[1][0] = mouseevent.xdata
-            self.span.xy[4][0] = mouseevent.xdata
-            self.draw()
-
-    def on_span_right_motion(self, mouseevent):
-        """Update data of span selector tool on right movement of mouse and
-        redraw.
-
-        """
-        if mouseevent.xdata > self.span.xy[0][0]:
-            self.span.xy[2][0] = mouseevent.xdata
-            self.span.xy[3][0] = mouseevent.xdata
-            self.draw()
-
-    def on_span_resized(self, mouseevent):
-        """Redraw the red rectangle to currently selected span."""
-        self.figure.canvas.mpl_disconnect(self.cid[1])
-        self.figure.canvas.mpl_disconnect(self.cid[2])
-        self.cid[0] = self.figure.canvas.mpl_connect('pick_event', 
-            self.on_span_pick)
-        self.span.set_edgecolor('r')
-        self.span.set_linewidth(1)
-        self.draw()
-
-    def on_shell_pick(self, event):
-        """Highlight the shell that was picked."""
-        self.highlight_shell(event.artist.index)
-
-    def highlight_shell(self, index):
-        """Change edgecolor of highlighted shell."""
-        self.parent.tableview.selectRow(index)
-        for i in range(len(self.parent.shells)):
-            if i != index and i != index + 1:
-                self.parent.shells[i].set_edgecolor('k')
-            else:
-                self.parent.shells[i].set_edgecolor('w')
-        self.draw()
-
-    def shell_picker(self, shell, mouseevent):
-        """Enable picking shells in the shell plot."""
-        if mouseevent.xdata is None:
-            return False, dict()
-        mouse_r2 = mouseevent.xdata ** 2 + mouseevent.ydata ** 2
-        if shell.r_inner ** 2 < mouse_r2 < shell.r_outer ** 2:
-            return True, dict()
-        return False, dict()
-
-    def span_picker(self, span, mouseevent, tolerance=5):
-        """Detect mouseclicks inside tolerance region of the span selector
-        tool and pick it.
-
-        """
-        left = float(span.xy[0][0])
-        right = float(span.xy[2][0])
-        tolerance = span.axes.transData.inverted().transform((tolerance, 0)
-            )[0] - span.axes.transData.inverted().transform((0, 0))[0]
-        event_attributes = {'edge': None}
-        if mouseevent.xdata is None:
-            return False, event_attributes
-        if left - tolerance <= mouseevent.xdata <= left + tolerance:
-            event_attributes['edge'] = 'left'
-            return True, event_attributes
-        elif right - tolerance <= mouseevent.xdata <= right + tolerance:
-            event_attributes['edge'] = 'right'
-            return True, event_attributes
-        return False, event_attributes
-
-class Shell(matplotlib.patches.Wedge):
-    """A data holder to store measurements of shells that will be drawn in 
-    the plot.
+class LineInteractionTables(QtGui.QWidget):
+    """Widget to hold the line interaction tables used by 
+    LineInfo which in turn is used by spectrum widget.
 
     """
-    def __init__(self, index, center, r_inner, r_outer, **kwargs):
-        super(Shell, self).__init__(center, r_outer, 0, 90, 
-            width=r_outer - r_inner, **kwargs)
-        self.index = index
-        self.center = center
-        self.r_outer = r_outer
-        self.r_inner = r_inner
-        self.width = r_outer - r_inner
+
+    def __init__(self, line_interaction_analysis, atom_data, description):
+        """Create the widget and set data."""
+        super(LineInteractionTables, self).__init__()
+        self.text_description = QtGui.QLabel(str(description))
+        self.species_table = QtGui.QTableView()
+        self.transitions_table = QtGui.QTableView()
+        self.layout = QtGui.QHBoxLayout()
+        self.line_interaction_analysis = line_interaction_analysis
+        self.atom_data = atom_data
+        line_interaction_species_group = \
+        line_interaction_analysis.last_line_in.groupby(['atomic_number', 
+            'ion_number'])
+        self.species_selected = sorted(
+            line_interaction_species_group.groups.keys())
+        species_symbols = [util.species_tuple_to_string(item, 
+            atom_data) for item in self.species_selected]
+        species_table_model = SimpleTableModel([species_symbols, ['Species']])
+        species_abundances = (
+            line_interaction_species_group.wavelength.count().astype(float) /
+            line_interaction_analysis.last_line_in.wavelength.count()).astype(float).tolist()
+        species_abundances = map(float, species_abundances)
+        species_table_model.addData(species_abundances)
+        self.species_table.setModel(species_table_model)
+
+        line_interaction_species_group.wavelength.count()
+        self.layout.addWidget(self.text_description)
+        self.layout.addWidget(self.species_table)
+        self.species_table.connect(self.species_table.verticalHeader(), 
+            QtCore.SIGNAL('sectionClicked(int)'), self.on_species_clicked)
+        self.layout.addWidget(self.transitions_table)
+
+        self.setLayout(self.layout)
+        self.show()
+
+    def on_species_clicked(self, index):
+        """"""
+        current_species = self.species_selected[index]
+        last_line_in = self.line_interaction_analysis.last_line_in
+        last_line_out = self.line_interaction_analysis.last_line_out
+
+        last_line_in_filter = (last_line_in.atomic_number == current_species[0]).values & \
+                              (last_line_in.ion_number == current_species[1]).values
+
+        current_last_line_in = last_line_in[last_line_in_filter].reset_index()
+        current_last_line_out = last_line_out[last_line_in_filter].reset_index()
+
+        current_last_line_in['line_id_out'] = current_last_line_out['line_id']
+
+
+        last_line_in_string = []
+        last_line_count = []
+        grouped_line_interactions = current_last_line_in.groupby(['line_id', 
+            'line_id_out'])
+        exc_deexc_string = 'exc. %d-%d (%.2f A) de-exc. %d-%d (%.2f A)'
+
+        for line_id, row in grouped_line_interactions.wavelength.count().iteritems():
+            current_line_in = self.atom_data.lines.ix[line_id[0]]
+            current_line_out = self.atom_data.lines.ix[line_id[1]]
+            last_line_in_string.append(exc_deexc_string % (
+                current_line_in['level_number_lower'],
+               current_line_in['level_number_upper'],
+               current_line_in['wavelength'],
+               current_line_out['level_number_upper'],
+               current_line_out['level_number_lower'],
+               current_line_out['wavelength']))
+            last_line_count.append(int(row))
+
+
+        last_line_in_model = SimpleTableModel([last_line_in_string, [
+            'Num. pkts %d' % current_last_line_in.wavelength.count()]])
+        last_line_in_model.addData(last_line_count)
+        self.transitions_table.setModel(last_line_in_model)
+
+class Tardis(QtGui.QMainWindow):
+    """Create the top level window for the GUI and wait for call to 
+    display data.
+
+    """
+    
+    def __init__(self, tablemodel, config=None, atom_data=None, parent=None):
+        """Create the top level window and all widgets it contains.
+
+        When called with no arguments it initializes the GUI in passive
+        mode. When a yaml config file and atom data are provided the 
+        GUI starts in the active mode.
+
+        Parameters
+        ---------
+            parent: None
+                Set to None by default and shouldn't be changed unless 
+                you are developing something new.
+            config: string 
+                yaml file with configuration information for TARDIS.
+            atom_data: string
+                hdf file that has the atom data.
+
+        Raises
+        ------
+            TemporarilyUnavaliable
+                Raised when an attempt is made to start the active mode. 
+                This will be removed when active mode is developed.
+
+        """
+
+        #assumes that qt has already been initialized by starting IPython 
+        #with the flag "--pylab=qt"
+        # app = QtCore.QCoreApplication.instance()
+        # if app is None:
+        #     app = QtGui.QApplication([])
+        # try:
+        #     from IPython.lib.guisupport import start_event_loop_qt4
+        #     start_event_loop_qt4(app)
+        # except ImportError:
+        #     app.exec_()
+
+        QtGui.QMainWindow.__init__(self, parent)
+
+        #path to icons folder
+        self.path = os.path.join(tardis.__path__[0],'gui','images')  
+
+        #Check if configuration file was provided
+        self.mode = 'passive'
+        if config is not None:
+            self.mode = 'active'
+
+        #Statusbar
+        statusbr = self.statusBar()
+        lblstr = '<font color="red"><b>Calculation did not converge</b></font>'
+        self.successLabel = QtGui.QLabel(lblstr)
+        self.successLabel.setFrameStyle(QtGui.QFrame.StyledPanel |
+            QtGui.QFrame.Sunken)
+        statusbr.addPermanentWidget(self.successLabel)
+        self.modeLabel = QtGui.QLabel('Passive mode')
+        statusbr.addPermanentWidget(self.modeLabel)
+        statusbr.showMessage(self.mode, 5000)
+        statusbr.showMessage("Ready", 5000) 
+
+        #Actions
+        quitAction = QtGui.QAction("&Quit", self)
+        quitAction.setIcon(QtGui.QIcon(os.path.join(self.path, 
+            'closeicon.png')))
+        quitAction.triggered.connect(self.close)
+        
+        self.viewMdv = QtGui.QAction("View &Model", self)
+        self.viewMdv.setIcon(QtGui.QIcon(os.path.join(self.path,
+            'mdvswitch.png')))
+        self.viewMdv.setCheckable(True)
+        self.viewMdv.setChecked(True)
+        self.viewMdv.setEnabled(False)
+        self.viewMdv.triggered.connect(self.switchToMdv)
+        
+        self.viewForm = QtGui.QAction("&Edit Model", self)
+        self.viewForm.setIcon(QtGui.QIcon(os.path.join(self.path,
+            'formswitch.png')))
+        self.viewForm.setCheckable(True)
+        self.viewForm.setEnabled(False)
+        self.viewForm.triggered.connect(self.switchToForm)
+
+        #Menubar
+        self.fileMenu = self.menuBar().addMenu("&File")
+        self.fileMenu.addAction(quitAction)
+        self.viewMenu = self.menuBar().addMenu("&View")
+        self.viewMenu.addAction(self.viewMdv)
+        self.viewMenu.addAction(self.viewForm)
+        self.helpMenu = self.menuBar().addMenu("&Help")
+
+        #Toolbar
+        fileToolbar = self.addToolBar("File")
+        fileToolbar.setObjectName("FileToolBar")  
+        fileToolbar.addAction(quitAction)
+
+        viewToolbar = self.addToolBar("View")
+        viewToolbar.setObjectName("ViewToolBar")
+        viewToolbar.addAction(self.viewMdv)
+        viewToolbar.addAction(self.viewForm)
+
+        #Central Widget
+        self.stackedWidget = QtGui.QStackedWidget()
+        self.mdv = ModelViewer(tablemodel) 
+        self.stackedWidget.addWidget(self.mdv)
+        
+        #In case of active mode
+        if self.mode == 'active':
+            #Disabled currently
+            # self.formWidget = ConfigEditor(config)
+            # #scrollarea
+            # scrollarea = QtGui.QScrollArea()
+            # scrollarea.setWidget(self.formWidget)
+            # self.stackedWidget.addWidget(scrollarea)
+            # self.viewForm.setEnabled(True)
+            # self.viewMdv.setEnabled(True)
+            # model = run_tardis(config, atom_data)
+            # self.show_model(model)
+            raise TemporarilyUnavaliable("The active mode is under" 
+                "development. Please use the passive mode for now.")
+
+        self.setCentralWidget(self.stackedWidget)
+
+    def show_model(self, model=None):
+        """Set the provided model into the GUI and show the main window.
+
+        Parameters
+        ----------
+        model: TARDIS model object 
+            A keyword argument that takes the tardis model object.
+
+        """
+        if model:
+            self.mdv.change_model(model)
+        if model.converged:
+            self.successLabel.setText('<font color="green">converged</font>')
+        if self.mode == 'active':
+            self.modeLabel.setText('Active Mode')
+
+        self.mdv.fillOutputLabel()
+        self.mdv.tableview.setModel(self.mdv.tablemodel)
+        self.mdv.plot_model()
+        self.mdv.plot_spectrum()
+        self.showMaximized()
+        
+
+    def switchToMdv(self):
+        """Switch the cental stacked widget to show the modelviewer."""
+        self.stackedWidget.setCurrentIndex(0)
+        self.viewForm.setChecked(False)
+
+    def switchToForm(self):
+        """Switch the cental stacked widget to show the ConfigEditor."""
+        self.stackedWidget.setCurrentIndex(1)
+        self.viewMdv.setChecked(False)
+
+class TemporarilyUnavaliable(Exception):
+    """Exception raised when creation of active mode of tardis is attempted."""
+    
+    def __init__(self, value):
+        self.value = value
+    
+    def __str__(self):
+        return repr(self.value)
